unixsock: add progressive timeout
Currently timout is hardcoded to 60sec which is not flexible because
if unix socket is used for communication with VM instances
(such as qemu,parallels,vmware and etc) hypervisor tend to restart
unix server during VM's reboot. Due to hardcoded timout reconnect
will happen too late and many important data will be lost.
Let's use progressive timout in range [1, 60] as we currently use for
process and telnet servers.

Signed-off-by: Dmitry Monakhov <dmonakhov@openvz.org>
Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>