unixsock: require min conn len before delay reset
Once a unixsock obj transitions to the UP state, require the new
connection to be up for a minimum length of time before resetting
the reconnect delay.  This protects against spinning on reconnects
when the connection is immediately terminated.

Other objects using an exponential reconnect delay require the
connection to be up for at least the minimum timeout.  This doesn't
make sense for a unixsock obj which now has a 1-second minimum timeout.
Instead, unixsock uses the new MIN_CONNECT_SECS which defines the
minimum number of seconds that a newly-connected object must be up
before resetting its reconnect delay back to the minimum.