set max on ipmiconsole_engine_init() num_threads
ipmiconsole_engine_init() returns EINVAL if more than
IPMICONSOLE_THREAD_COUNT_MAX (32) threads are requested.

That limit is exceeded if more than 4096 IPMI SOL consoles are defined:

  num_threads = ((num_consoles - 1) / IPMI_ENGINE_CONSOLES_PER_THREAD) + 1;

As such, set a maximum bound on the number of threads passed to
ipmiconsole_engine_init() in order to ensure this upper limit is
not exceeded.

Fixes issue 15.