Update buffer constants
This revisits the constants used for buffers by the client and server.
Essentially, the size of the circular buffers for objects are doubled
which should reduce the likelihood of characters being dropped before
they are written out.

MAX_BUF_SIZE is used in the client and server for different purposes.
The client keeps MAX_BUF_SIZE for use as a generic buffer size,
and its value is reduced to 4096.  The server replaces this with
OBJ_BUF_SIZE for use as the size of the object buffers, and that
value is increased to 16384.

CONMAN_REPLAY_LEN is renamed to LOG_REPLAY_LEN.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>