Add conmand "-P" cmdline opt for setting pidfile
conmand typically runs as a background process spawned via a
double-fork() as many traditional UNIX daemons are wont to do.

With systemd, Type=forking services are recommended to use the
PIDFile option so systemd can identify the main process of the daemon.
But conmand only supports specifying a PID file in its config file
(server pidfile="<file>"), and this value is unset by default.

This commit adds the "-P <file>" command-line option to conmand
in order to specify the daemon's PID file.  This allows the systemd
service file to specify a PID file without relying on the conmand
config file to match.  The command-line option will take precedence
over the config file setting.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>