Convert RPM specfile from sysvinit to systemd
This commit converts the RPM specfile from sysvinit to systemd.
The requisite Requires & BuildRequires tags for systemd are added.
References to chkconfig and the sysvinit script are replaced with
corresponding helper macros to handle systemd scriptlet operations.
The sysvinit script and sysconfig file are replaced with the systemd
service unit configuration file.

Makefile.in is updated to conditionally install the systemd service
file if the directory "$(prefix)/lib/systemd/system" exists.  Ideally,
there should probably be a configure option to select between sysvinit
and systemd installations.

The systemd service file is configured to start conmand as a
"forking" service; when run in this manner, conmand honors the
"server logfile='<file>'" directive to write its log messages to the
specified file.  Note that when conmand is run in the foreground,
log messages are instead written to stderr.  The limit for the number
of open files (LimitNOFILE) is set to "infinity" in order to specify
no limit, thereby allowing this limit to be controlled by the "server
nofile=<int>" directive in conman.conf.

References:
- https://fedoraproject.org/wiki/Packaging:Systemd
- https://fedoraproject.org/wiki/Packaging:ScriptletSnippets

Tested on:
- CentOS 7
- Fedora 23

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>