Update RPM specfile
Change the Source0 tag to use the full URL to the distributed tarball.

Remove the BuildRoot tag since it is ignored as of Fedora 18.

Remove %defattr at the beginning of each %files section since these
permissions are the default as of rpm 4.4.

Change the tag for COPYING from %doc to %license.  Also add a
conditional test for _licensedir to define %license if needed.

References:
- https://fedoraproject.org/wiki/How_to_create_an_RPM_package
- https://fedoraproject.org/wiki/Packaging:Guidelines
- https://fedoraproject.org/wiki/Packaging:LicensingGuidelines

Tested on:
- CentOS 7
- Fedora 23

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>