Remove special case for close after failed connect
This removes code for a special case that occurred on Tru64 many
years back where the close() of a socket that failed to connect()
would return an error of "invalid argument".  If this special
case still persists on current systems, it should be handled in
disconnect_telnet_obj() when the fd is closed.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>