Switch to event-based computation of tpoll fd-set
The main loop in conmand is mux_io().  It computes the set of
file descriptors to be poll()d, invokes poll() (via the tpoll()
abstraction), and processes activity on the file descriptors of
interest.  This is inefficient for primarily two reasons.  First,
recomputing the set of file descriptors on each iteration requires
a full list traversal of all objects as well as complicated logic
to determine which descriptors to poll for which activity.  Second,
poll() does not scale well since it requires a list traversal to
process the results.

This commit removes the recomputation of the file descriptor set.
Instead, modifications to this set are triggered by events (e.g.,
when an object connects or transitions to its UP state, when new data
is added, when all data is written, etc.).

This commit does not remove the list traversal before tpoll().
That remains solely to process the 'gotReset' flag signaling that
the associated console should be "reset".  It will be restructured
in a later commit in order to remove that first list traversal.

The 1000-ms timeout value to tpoll() is removed.  This timeout value
was necessary for accepting new clients.  Upon a successful accept(),
a thread is created for process_client().  For "connect" and "monitor"
commands, a new object is then created via create_client_obj().
That routine adds a new descriptor to the tpoll() file descriptor set.
POLLIN activity for this new descriptor would be added on the next
iteration of the mux_io() loop when the file descriptor set is
recomputed.  Without the timeout value, tpoll() could block on the
set of old descriptors and never be awoken to process activity on the
new client.  However, with the event-based changes to the tpoll() file
descriptor set, create_client_obj() calls tpoll_set() which can unblock
tpoll() when necessary.  Consequently, the timeout to tpoll() is no
longer needed; it is now set to -1 to block until I/O is available,
a timer has expired, or the file descriptor set has changed.

The "priming optimization" has been removed from read_from_obj().
This optimization dates back to when mux_io() was implemented
with select().  Since select() modifies the passed file descriptor
set, write activity for an object could be "primed" by setting
the write-activity bit when adding data to an object's buffer;
this allowed new data to be written out to those objects not yet
traversed during the current list iteration, thereby reducing latency.
However, poll() uses separate bitmasks for the input value (events)
and output result (revents).  As such, priming has no effect here.
This code was erroneously carried over when the tpoll() abstraction
was introduced.  And now its watch has ended.

Finally, the tpoll_t argument was removed from read_from_obj().
This variable is already passed as a global (tp_global) to several
files in order to allow timers to be set.  Passing it as a parameter
to this function serves no useful purpose.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>