Revisit errors when closing obj file descriptors
The following error was observed on our jade cluster:

> INFO:  Console [jade1326] connected to <pjade1326>
> ...
> INFO:  Unable to read from [jade1326]: Bad file descriptor
> ERROR: Unable to close object [jade1326]: Bad file descriptor

jade1326 is an IPMI console that was in a connected state with a
valid file descriptor.  Somehow the file descriptor went bad and the
sequence above unfolded as follows:

  read_from_obj() invoked
    read() returned -1 with EBADF
    log_msg("Unable to read from [jade1326]: Bad file descriptor")
    shutdown_obj() invoked
      obj->fd > 0
      close() returned -1 with EBADF
      log_err("Unable to close object [jade1326]: Bad file descriptor")
        exit(1)

The failed close() in shutdown_obj() should not be fatal since a new
connection for that console will be attempted.

This commit attempts to make the code more consistent with how errors
are handled when closing an object's file descriptor.  When these
errors occur, a warning message should be logged and processing
should continue.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>