Increase MAX_SOCK_LINE buffer constant
This increases the MAX_SOCK_LINE buffer constant from 32KB to 128KB.
This constant is for sizing the buffers used during the client-server
protocol.

The problem is when conmand is managing several thousand consoles.
The client-server protocol is not particularly efficient.  If the
client queries the server for a list of all consoles, the server
can overrun its response buffer at which point it will terminate
the connection:

> $ conman -q
> ERROR: Connection terminated by <localhost:7890>.

Our jade cluster has 2303 consoles and requires this constant to be
at least 42KB.  By increasing it to 128KB, clients should be able to
query up to ~8000 consoles (less if the console names are longer).

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>