Remove mux_io() list traversal for console reset
This commit revamps how a console reset is processed.

The console reset escape sequence is first handled by
process_client_escapes() which then calls perform_reset().  These
functions have a pointer to the client object but do not have access
to the conf struct in which the ResetCmd string resides.  By setting
the gotReset flag in the console's base_obj struct, the reset could
be performed when control returned to mux_io() which had access to
the conf struct.

Commit 95e57b4 began restructuring mux_io() which would fully
traverse the object list both before and after each call to poll().
The first list traversal would compute the file descriptor set, as
well as process the gotReset flag indicating whether the associated
console should be reset.  The second list traversal would process the
results of the poll().  After switching to an event-based computation
of the tpoll file descriptor set, the list traversal before poll()
served only to process the gotReset flag.

This commit begins by replacing the gotReset flag in the base_obj
struct with resetCmdRef, resetCmdPid, and resetCmdTimer.  resetCmdRef
points to the ResetCmd string stored in the conf struct.  This is
set for each console object during the object list traversal in
open_objs().  Since this pointer is now stored in the console object,
it can be accessed from perform_reset() via the given client object.

The mux_io() object list traversal for servicing gotReset is removed.
The functionality in reset_console() for performing the fork-and-exec
of the ResetCmd is moved into perform_reset(), and kill_console_reset()
is moved and renamed to kill_reset_cmd().

When a ResetCmd is forked by perform_reset(), a timer is created to
ensure the new process terminates within RESET_CMD_TIMEOUT seconds.
At expiration, this timer checks via kill(0) to see if the process is
still running, and if so, issues a kill(9).  If the ResetCmd process
exits before then, its status is reaped by sig_chld_handler() but
that signal handler does not attempt to lookup the associated timer
and cancel it.  As such, if an unrelated process should re-use this
pid within the RESET_CMD_TIMEOUT interval, it would be erroneously
killed by the daemon.  Note that this existing behavior should ideally
be corrected at some point.

Also note that a client cannot issue another reset until the former
one completes.  Since sig_chld_handler() does not update the associated
console object when the ResetCmd process terminates, the client must
wait RESET_CMD_TIMEOUT seconds for the timer to clear the console
object's resetCmdPid before another reset can be performed.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>