Allow write_notify_msg() to write partial messages
This commit revamps the write_notify_msg() function that writes a
notification message to the daemon logfile and all attached readers
and writers of a given console.

If the buffer was exceeded during message construction, the function
would immediately return to the caller instead of writing a truncated
log message.  This commit changes that behavior to log as much of
the message that fits in the buffer.

Also, this commit removes the short time string being appended to
the console notification message.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>