Replace write_to_obj() write with writev
write_to_obj() is responsible for writing data from the object's
circular buffer out to its file descriptor.  When data had wrapped
around the circular buffer, only the first chunk of data would be
written out; the remaining wrapped-around data would be written during
the next invocation of the routine for this object.

This simplified the code for handling a wrap-around, but added an
additional write() system call and increased latency by delaying this
subsequent write until the next iteration of the mux_io() loop.  If the
daemon was under heavy load (particularly when managing thousands of
consoles), this behavior increased the likelihood that data could be
overwritten in the circular buffer before being written out to its
file descriptor.

This commit replaces the write() with a writev(), thereby allowing any
data that has wrapped around to be processed in the same system call.

This commit also removes a check that ensures the console connection
state is up and clears the circular buffer if the connection state
has gone down.  This logic does not belong in the call path for
every write, especially now that the tpoll file descriptor set is
event-based.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>