Replace out_of_memory() macro with function call
Both lex.c and list.c support a custom out_of_memory() handler.
This was specified by a preprocessor #define set via AH_BOTTOM
in configure.ac (which caused the #define to appear in config.h).
However, gcc 4.9.2 on Debian 8 was generating several warnings like
the following:

> gcc -DHAVE_CONFIG_H -I.  -g -Wall  -c common.c
> In file included from util-str.h:29:0,
>                  from common.c:36:
> common.c: In function 'create_req':
> ./config.h:175:53: warning: right-hand operand of comma expression has
>       no effect [-Wunused-value]
>  #define out_of_memory() (log_err(0, "Out of memory"), (void *) NULL)
>                                                      ^
> common.c:88:9: note: in expansion of macro 'out_of_memory'
>          out_of_memory();
>          ^

In order to suppress this warning, the preprocessor macro in
configure.ac is replaced with the function already defined in util.c.
Furthermore, WITH_OOMF is defined via the Makefile DEFS variable in
order to use this out_of_memory() function in lex.c and list.c.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>