Prevent mux_io from spinning when client suspended
When a client suspends, the server sets the associated gotSuspend
flag.  But POLLOUT for the associated file descriptor will still
be set whenever data is written into the client object's circular
buffer, and data in the circular buffer cannot be written to the
client while it is suspended.  This causes activity on the client's
file descriptor that cannot be cleared until the client resumes.
Consequently, mux_io() needlessly spins.

This bug was introduced with the switch to an event-based computation
of the tpoll file descriptor set in commit 95e57b4.

This commit clears POLLOUT when the client suspends, and sets POLLOUT
when the client resumes.  In addition, it prevents POLLOUT from being
set while the client is suspended.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>