Change tpoll_set to update max_fd only for new fd
In tpoll_set(), max_fd is potentially updated if the "events" field in
the pollfd struct has changed (i.e., if it differs from "events_new").
But max_fd only needs to be potentially updated when a new fd is set.

This commit moves the potential update to max_fd so it only happens
when a new fd is set.  This avoids needlessly checking max_fd when
the events for an active fd are merely changed.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>