Fix read_from_obj comment wrt pending telnet objs
Before the switch to an event-based computation of the tpoll file
descriptor set in commit 95e57b4, mux_io() would check if the given
file descriptor was associated with a telnet connection in the PENDING
state before processing any read or write activity for that file
descriptor.  This is alluded to in read_from_obj() where a comment
states that a PENDING telnet obj will be forced into either the UP
or DOWN state via open_telnet_obj() before read_from_obj() is called.

When a non-blocking connect completes, the file descriptor becomes
writable; when connection establishment fails, it becomes both readable
and writable.  As such, commit 95e57b4 moves the completion of a
PENDING telnet connection from mux_io() to write_to_obj() (which
processes write activity).  Thus, the comment in read_from_obj()
is incorrect since the routine can be called with a telnet obj in
the PENDING state.

This commit corrects the comment in read_from_obj().

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>