Add test console device
This commit adds a test console device to aid in development and
performance testing.  It simulates reads in a manner reminiscent of
lptest, writing them to any attached clients and logfiles.  Any data
received (from clients) is written to /dev/null.

This device is specified with a console directive having dev="test:".
The "testopts" string specifies additional options.  It consists of
comma-delimited tokens, each of the form "X:VALUE" where "X" is a
single-character key tag.  The following tags are supported:

  B - the number of bytes to output on each simulated read
      (defaults to 1024)

  M - the maximum number of milliseconds between reads
      (defaults to 100)

  N - the minimum number of milliseconds between reads
      (defaults to the maximum which results in a constant delay)

  P - the probability that the read will occur
      (defaults to 100)

A global testopts setting can also be specified.

Examples:

  global testopts="b:8191,m:50"
  console name="foo" dev="test:" testopts="b:1,m:500,n:50,p:75"
  console name="bar" dev="test:" log="%N.log" testopts="M:100,B:80"

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>