Revamp EOF processing in read_from_obj
When an EOF is read() from the object's file descriptor, the gotEOF
flag is set to prevent new data from being added to the object's
buffer while the existing data is written out to the file descriptor.
Once gotEOF has been set, read (POLLIN) activity should no longer be
polled on this file descriptor -- but this is not the case.

The failure to clear POLLIN here was introduced with the switch
to an event-based computation of the tpoll file descriptor set in
commit 95e57b4.

Once gotEOF is set, a subsequent read() on this file descriptor
will trigger shutdown_obj().  This now happens each time a client
disconnects which results in the following log message:

  Attempted to read from [foo@localhost:56789] after EOF

This commit fixes this behavior by clearing POLLIN once gotEOF is set.
In addition, it immediately triggers shutdown_obj() if the object's
buffer is already empty.

The check for gotEOF has been removed from the top of read_from_obj()
since it should no longer be necessary with POLLIN having been cleared.
Instead of testing for it on every invocation, a check for this flag
(i.e., testing for multiple EOFs) has been added to the code block
handling just the EOF.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>