Log num bytes dropped when obj buffer is flushed
When writing data out to an object's file descriptor in write_to_obj(),
a writev() failure will trigger shutdown_obj().  And shutdown_obj()
will flush all buffered data.

This commit adds a log message to shutdown_obj() with the number of
bytes of unwritten data being dropped.  This count is performed with
the new function num_bytes_buffered().

In destroy_obj(), the log message added in commit 49a846b to warn of
unflushed data is updated to use this function.

In write_obj_data(), the computation of "avail" (the number of bytes
available in that object's buffer before unwritten data is overwritten)
is also updated to use this function.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>