Remove write_obj_data check for console state
This commit removes a multi-condition check for ensuring the obj
is in an "up" state with a valid fd before writing data into its
circular-buffer.  This logic does not belong in the call path for
every write.  Data can be written into an obj's buffer as long as the
gotEOF flag has not been set.  And the fd is validated at the top of
write_to_obj() before the writev() occurs.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>