Add log msg when client is suspended or resumed
When the client is suspended via the "&Z" escape sequence, it first
signals the server before throwing a SIGTSTP.  The server processes
this escape sequence in perform_suspend().  This routine toggles a
flag to prevent sending data to the client while it is suspended;
it also writes a debug message.

This debug message is conditionally compiled via "--enable-debug".
If debugging is not enabled, gcc 4.9.2 on Debian 8.6 generates the
following compiler warning:

> server-esc.c: In function 'perform_suspend':
> server-esc.c:517:9: warning: variable 'gotSuspend' set but not used
>       [-Wunused-but-set-variable]

This commit replaces the debug message with an informational log
message that is not conditionally compiled.  It also removes the
"gotSuspend" variable.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>