Revisit log messages in server-unixsock.c
The log message for a failed S_ISSOCK is changed to have a LOG_NOTICE
priority level.  Additionally, its final arg of "strerror(errno)"
is removed since it is not used in the format string.

The assert() for a failed strlcpy() when exceeding "saddr.sun_path"
is replaced with a runtime check that logs a message at the LOG_NOTICE
priority level and fails the connection attempt.  This also resolves
an unused-but-set-variable compiler warning (when debugging was not
enabled) for the return value of strlcpy() which was only being used
in the assert().

The fatal log_err() for a failed socket() is changed to log a message
and failed connection attempt.

Finally, the wording of log messages is tweaked for consistency.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>