Setup Travis CI builds
This commit enables Travis CI continuous integration builds at
<https://travis-ci.org/>.

A build matrix is defined to create debug and production builds using
gcc and clang compilers.

The "-Werror" and "-pedantic-errors" CFLAGS options are specified
for all builds in order to change warnings into errors.

Note that the "--with-freeipmi" configure option is not enabled.
The standard Travis CI virtualization environment is based on Ubuntu
12.04, and its version of FreeIPMI is too old to work with ConMan;
also, libfreeipmi-dev and libipmiconsole-dev are not on the list of
approved packages.  Installing a current version of FreeIPMI for the
Travis builds is TBD.

colormake is used to colorize the output of the Travis CI job log.

Reference:
- https://docs.travis-ci.com/user/getting-started/

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>