Clear gotEOF during shutdown_obj
This commit clears the gotEOF flag during shutdown_obj() when the
buffer is flushed.  Normally this is not an issue since the gotEOF
flag is cleared when the object reconnects.  But if gotEOF is not
reset during shutdown, the object would be prevented from having data
written in to its buffer before it reconnected.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>