Fix resource leak of sd in connect_to_server
In connect_to_server(), a file descriptor is obtained for an unbound
socket.  If the subsequent call to host_name_to_addr4() or connect()
fails, the function will set errnum & errmsg and return -1 without
closing the socket descriptor thereby causing a resource leak --
although the function will then immediately return to main(), call
display_error(), and terminate.

This commit fixes the resource leak of the socket descriptor in
connect_to_server().  The return value of close() is discarded since
it is in the error handling code.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>
Coverity ID 1361793
Coverity ID 1367924