Fix resource leak for unixsock in process_console
When a unixsock console is created in process_console(), memory
is allocated in is_unixsock_dev() for a string containing the
absolute pathname of the Unix domain socket.  A pointer to this
string is stored in the variable "path".  This "path" is passed to
create_unixsock_obj() to create the console object, but the memory
is never free()d afterwards.

This commit fixes the resource leak of the unixsock path string in
process_console().  This bug originated in commit 879a641.  It is
a cut & paste error where the code from the telnet object creation
block was copied and tweaked for unixsock.  Instead of freeing the
memory for "path", the code was left unchanged to free the memory for
"host"; erroneously freeing this "host" did not cause an error since
the variable had been initialized to NULL.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>
Coverity ID 1361789