Fix negative return passed to dup2 in perform_reset
perform_reset() does a fork & exec of /bin/sh to run the reset
command for each console being reset.  Within the child process,
/dev/null is open()d and dup2()d onto stdin, stdout, and stderr;
this ensures the reset command won't unexpectedly fail if it performs
I/O on these descriptors.  However, the return value of open() is not
checked since the error cannot be logged from the child.  This allows
a negative value to be passed as the source file descriptor to dup2().

This commit open()s /dev/null in the parent instead of the child
in order to log the error if needed.  If the open() is successful,
the child will dup2() this descriptor onto stdin, stdout, and stderr;
otherwise, it will just close() them and continue with the reset.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>
Coverity ID 1361787