Remove logically dead code in process_telnet_cmd
In process_telnet_cmd(), the WONT case of the switch statement logs a
message noting the received telnet command & option.  A conditional
expression is used to display the option string; however, the false
condition for TELOPT_OK() will never be executed since this code
block is guarded by an if-statement testing for one of three valid
telnet options.

This commit removes the logically dead code in the conditional
expression of the log message.  This code was presumably copied from
the log message statement in the default case which needs to validate
the telnet option.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>
Coverity ID 1361786