Replace LEX_UNTOK with LEX_TOK2STR
This commit replaces the remaining uses of the LEX_UNTOK(tok) macro
with the new LEX_TOK2STR(tokstrs,tok) macro.  LEX_UNTOK is removed.
LEX_TOK2STR differs from LEX_UNTOK in that it returns a string from
the specified tokstrs[] array.  No bounds-checking is performed,
so LEX_TOK2STR should only be used when the tok enum value is known
to correspond to a string in the tokstrs[] array.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>