Ensure snprintf size_t arg is not a negative value
Several calls to snprintf() and vsnprintf() do not ensure that the
value for the unsigned integer "size" argument (of type size_t)
is non-negative.  In several cases, an integer variable (signed)
is used to store the buffer length.  This variable is used for the
"size" argument in the call to snprintf().  The signed integer gets
promoted to an unsigned integer.  If its value was negative, it will
be treated as a very large positive integer by the snprintf() "size"
argument and potentially result in a buffer overrun.

This commit fixes these instances of snprintf() and vsnprintf()
by adding guards to ensure the "size" argument is non-negative.
It also ensures the "str" argument for the buffer address is non-NULL.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>