Revisit perform_log_replay
Coverity Scan flagged an out-of-bounds read in perform_log_replay().
Based on its analysis, it appears that Coverity is getting confused
by the bounds-checking performed when computing the number of bytes
to replay.  Its analysis increments "ptr" to an impossible value
which then results in the out-of-bounds read.

This commit adds a few assert()s and a check to ensure the number of
bytes to replay is not negative.  It also restructures some code for
readability.  But ultimately, the Coverity issue is being classified
as a false positive.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>
Coverity ID 1361791
Issue #24