Revisit append_format_string
append_format_string() appends a format-specified string to a
null-terminated buffer.  It then returns the new length of the
null-terminated string in this buffer (excluding the terminating null
byte), or -1 if truncation occurred.  But its return type is size_t
which is unsigned.  This commit begins by changing the return type
to an int.

Since this function is not declared static, the assert() that the
'dst' parameter be non-NULL is replaced with an if-guard.

The behavior of the function is changed slightly to ensure the 'dst'
array will always be null terminated as long as 'size' is non-zero.
Before this change, the function would return the 'dst' array unchanged
(with a return value of -1) if a null byte did not appear within the
first 'size' bytes of the array.  Now, the function will overwrite
the last byte with a null byte and still return -1.

Finally, the variables used to track the number of bytes in the
original string (num_orig) and the number of bytes remaining in the
array (num_left) are renamed, and their type is changed from int
to size_t.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>