Check all send_rsp retvals for buf overrun
send_rsp() builds up a response string in a local buffer 'buf' with
various calls to append_format_string(), strlcpy(), and snprintf().
Coverity Scan flagged an instance where a return value was overwritten
before it was used.

The return values from these functions are ignored (and overwritten)
until just before the response is sent to the client.  At that point,
the return value from the last write into buffer 'buf' is checked to
see if truncation has occurred.

This commit checks each of the return values for the routines writing
into the local buffer 'buf'.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>
Coverity ID 1361800