Fix perform_help_esc values overwritten before use
perform_help_esc() builds up a help message in a local buffer 'buf'
with several calls to append_format_string().  Coverity Scan flagged
several instances where the return value of append_format_string()
was being overwritten before it was used.

The return values of append_format_string() here are ignored except
for the last which is checked to see if truncation has occurred.
This is safe to do since append_format_string() will not write beyond
the specified length of the buffer.

This commit explicitly ignores the return values of all but
the last call to append_format_string() by casting them to void.
This strategy would not be particularly efficient if an early call
to append_format_string() was likely to exceed the buffer since each
subsequent call would pointlessly scan the entire buffer for the
terminating null character.  But the size of the local buffer 'buf'
is sufficient to ensure this does not happen.

Finally, the return value of the last call to append_format_string() is
used in place of strlen() to specify the amount of data to write in the
call to write_n(); this avoids an unnecessary traversal of the string.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>
Coverity ID 1361801