Fix IPMI connection failure log message of success
When an IPMI connection fails, an informational message is logged.
The reason for the failure is obtained from ipmiconsole_ctx_errnum()
which returns the errnum of the most recently recorded error for the
given IPMI context.  Unfortunately, it sometimes returns "success"
for a failed connection:

  INFO: Unable to connect to <127.0.0.1> via IPMI for [i]: success

This commit fixes the log message such that if the error cannot be
determined, the message does not indicate a failure due to success.

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>