If connect fails let other side accept connection and come back
The UNIX socket code uses inotify to get notified when a socket
it monitors appears. The inotify is triggered when the other side
does bind(). However, we are not able to connect until the other
side calls listen(). Therefore, we need to let the other side continue
if the connect fails and should come back at the earliest time to
attempt a connect() again.
For this we just reset the poll() delay.

Signed-off-by: Egbert Eich <eich@suse.com>