Move lib/ files to $pkgdatadir
The lib/ directory contains various Expect scripts for installation
in ${prefix}/lib/.  But /usr/lib/ is intended for libraries for the
binaries in /usr/bin/ and /usr/sbin/.  ${libdir} is for object files
and libraries of object code, not installed executables or scripts.
Consequently, the installation prefix for files in the lib/ directory
is being changed to ${pkgdatadir}, the directory for package-specific
"idiosyncratic read-only architecture-independent data".

conman.exp contains common routines used throughout the Expect scripts.
alpha.exp contains common routines for interacting with the Compaq
Alpha Linux cluster (furnace) we had back in 2001.  These are being
re-organized into a lib/ subdirectory.

For the scripts in lib/examples/, the source command path to conman.exp
and alpha.exp is being updated accordingly.

conman.conf.5.in is updated to use "@datadir@/@PACKAGE@/exec" since
AC_CONFIG_FILES does not substitute the automake variable @pkgdatadir@.

Reference:
- https://www.gnu.org/software/autoconf/manual/autoconf-2.63/html_node/Installation-Directory-Variables.html

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>