Substitute CONMAN_CONF, CONMAN_HOST, CONMAN_PORT
Substitute the variables CONMAN_CONF, CONMAN_HOST, and CONMAN_PORT.
These autoconf variables are only used in the manpage templates files.
CONMAN_CONF is based on ${sysconfdir}, and its substitution conforms
to the GNU coding standards for installation directory variables.

Note that the configure.ac values for these variables must be kept
in sync with the respective variables in common.h and server.h.

Reference:
- https://www.gnu.org/software/autoconf/manual/autoconf-2.63/html_node/Installation-Directory-Variables.html

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>