Expand SYSCONFDIR into string constant via stringizing
Expand SYSCONFDIR into a string constant via stringizing.  (Some older
documentation refers to it as stringification.)  This uses two levels
of preprocessor macros: STRINGIZE and QUOTE.

The macro argument for STRINGIZE is stringized so it is not
macro-expanded first.  But the macro argument for QUOTE is completely
macro-expanded before the macro QUOTE itself is expanded to use
STRINGIZE.  For example:

  #define QUOTE(s) STRINGIZE(s)
  #define STRINGIZE(s) #s
  #define SYSCONFDIR /etc
    (or add -DSYSCONFDIR='/etc' to CPPFLAGS)

  STRINGIZE(SYSCONFDIR)
  -> "SYSCONFDIR"

  QUOTE(SYSCONFDIR)
  -> QUOTE(/etc)
  -> STRINGIZE(/etc)
  -> "/etc"

For the CONMAN_CONF #define directive, the preprocessor expands
SYSCONFDIR and replaces the stringized argument with a string constant.
The C compiler then concatenates the two adjacent string constants
into a single string.

Reference:
- https://gcc.gnu.org/onlinedocs/cpp/Stringizing.html

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>