Add makefile dependency for SYSCONFDIR
Starting in commit 547db3c, the $(sysconfdir) makefile variable is
passed in conmand_CPPFLAGS via the SYSCONFDIR preprocessor macro.
That macro is used to define CONMAN_CONF in "server.h".

If the configure script is re-run to change the value of $(sysconfdir)
(either directly, or indirectly through $(prefix)) after the
executables have been initially built, the conmand binary is not
updated to reflect the change to CONMAN_CONF.

Add a makefile dependency on the Makefile itself for the
"server-conf.c" object file (conmand-server-conf.$(OBJEXT)) which
contains the CONMAN_CONF string.

Tested:
- Debian 8.11

Signed-off-by: Chris Dunlap <cdunlap@llnl.gov>